Added a forecasting method
These changes consist on adding a forecasting method in order to determine the reserves for future development years. 2 methods were applied to the studied data, the first one assumes that the last development year is 2015, and the second one takes into consideration that the assumption took in the first method is not a relevant one and then it bases its calculations on extrapolation of the link factors from a development year to another.